fix: joining blob path to PDP public URL
