+++ cmd/cli/serve/full.go
@@ -23,6 +23,7 @@ import (
 	"github.com/storacha/piri/pkg/config"
 	configapp "github.com/storacha/piri/pkg/config/app"
 	"github.com/storacha/piri/pkg/fx/app"
+	"github.com/storacha/piri/pkg/mahttp"
 )
 
 var (
@@ -379,11 +380,10 @@ func buildServicesConfig(cfg config.UCANServer, publicURL *url.URL) (configapp.S
 		if err != nil {
 			return configapp.ServicesConfig{}, fmt.Errorf("converting PDP URL to multiaddr: %w", err)
 		}
-		pieceAddr, err := multiaddr.NewMultiaddr("/http-path/" + url.PathEscape("piece/{blobCID}"))
+		blobAddr, err = mahttp.JoinPath(curioAddr, "piece/{blobCID}")
 		if err != nil {
-			return configapp.ServicesConfig{}, fmt.Errorf("creating piece multiaddr: %w", err)
+			return configapp.ServicesConfig{}, fmt.Errorf("joining blob path to PDP multiaddr: %w", err)
 		}
-		blobAddr = multiaddr.Join(curioAddr, pieceAddr)
 	}
 
 	return configapp.ServicesConfig{
