+++ pkg/service/publisher/publisher.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net/url"
 	"slices"
 	"sync"
 
@@ -28,6 +27,7 @@ import (
 	"github.com/storacha/go-ucanto/core/result"
 	"github.com/storacha/go-ucanto/core/result/ok"
 	"github.com/storacha/go-ucanto/principal"
+	"github.com/storacha/piri/pkg/mahttp"
 
 	"github.com/storacha/go-libstoracha/advertisement"
 )
@@ -289,7 +289,7 @@ func New(
 func providerInfo(peerID peer.ID, publicAddr multiaddr.Multiaddr, blobAddr multiaddr.Multiaddr) (peer.AddrInfo, error) {
 	provider := peer.AddrInfo{ID: peerID}
 	if blobAddr == nil {
-		addr, err := joinHTTPPath(publicAddr, "blob/{blob}")
+		addr, err := mahttp.JoinPath(publicAddr, "blob/{blob}")
 		if err != nil {
 			return peer.AddrInfo{}, fmt.Errorf("joining blob pattern path to public multiaddr: %w", err)
 		}
@@ -297,52 +297,13 @@ func providerInfo(peerID peer.ID, publicAddr multiaddr.Multiaddr, blobAddr multi
 	}
 	provider.Addrs = append(provider.Addrs, blobAddr)
 
-	claimAddr, err := joinHTTPPath(publicAddr, "claim/{claim}")
+	claimAddr, err := mahttp.JoinPath(publicAddr, "claim/{claim}")
 	if err != nil {
 		return peer.AddrInfo{}, fmt.Errorf("joining claim pattern path to public multiaddr: %w", err)
 	}
 	provider.Addrs = append(provider.Addrs, claimAddr)
 
 	return provider, nil
-}
-
-// joinHTTPPath joins a HTTP path onto an existing multiaddr. If the multiaddr
-// includes the "http-path" protocol already, it is joined to the end of the
-// existing path. If the multiaddr does not contain the "http-path" protocol
-// then it is appended to the multiaddr, with the value specified.
-func joinHTTPPath(addr multiaddr.Multiaddr, path string) (multiaddr.Multiaddr, error) {
-	found := false
-	var out multiaddr.Multiaddr
-	for _, comp := range addr {
-		if comp.Code() == multiaddr.P_HTTP_PATH {
-			p, err := url.PathUnescape(comp.Value())
-			if err != nil {
-				return nil, err
-			}
-			u, err := url.Parse(p)
-			if err != nil {
-				return nil, err
-			}
-			p = u.JoinPath(path).Path
-			c, err := multiaddr.NewComponent("http-path", url.PathEscape(p))
-			if err != nil {
-				return nil, err
-			}
-			comp = *c
-			found = true
-		}
-		out = append(out, comp)
-	}
-
-	if !found {
-		c, err := multiaddr.NewComponent("http-path", url.PathEscape(path))
-		if err != nil {
-			return nil, err
-		}
-		out = append(out, *c)
-	}
-
-	return out, nil
 }
 
 func asCID(link ipld.Link) cid.Cid {
